+++ src/drivers/OPH_DUPLICATE_operator.c
@@ -673,7 +673,7 @@ int task_execute(oph_operator_struct * handle)
 
 	int l;
 
-	int num_threads = (oper_handle->nthread <= oper_handle->fragment_number ? oper_handle->nthread : oper_handle->fragment_number);
+	int num_threads = (oper_handle->nthread <= (unsigned int) oper_handle->fragment_number ? oper_handle->nthread : (unsigned int) oper_handle->fragment_number);
 	int res[num_threads];
 
 	//In multi-thread code mysql_library_init must be called before starting the threads
@@ -981,8 +981,8 @@ int task_destroy(oph_operator_struct * handle)
 		//Delete fragments
 		int num_threads =
 		    (((OPH_DUPLICATE_operator_handle *) handle->operator_handle)->nthread <=
-		     ((OPH_DUPLICATE_operator_handle *) handle->operator_handle)->fragment_number ? ((OPH_DUPLICATE_operator_handle *) handle->
-												     operator_handle)->nthread : ((OPH_DUPLICATE_operator_handle *) handle->operator_handle)->
+		     (unsigned int) ((OPH_DUPLICATE_operator_handle *) handle->operator_handle)->fragment_number ? ((OPH_DUPLICATE_operator_handle *) handle->
+												     operator_handle)->nthread : (unsigned int) ((OPH_DUPLICATE_operator_handle *) handle->operator_handle)->
 		     fragment_number);
 
 		if (((OPH_DUPLICATE_operator_handle *) handle->operator_handle)->fragment_id_start_position >= 0 || handle->proc_rank == 0) {
